UI/UX improvements by moving the getSortedPostsAsyncThunk call inside of the 'EDIT A POST' functionality
